+++ lib.h
